+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
-## [Unreleased]
+## [1.4.4] - 2019-04-15
 ### Fixed
 - Param `referrerId` in create wallet POST request
 
@@ -55,7 +55,8 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.4.3...HEAD
+[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.4.4...HEAD
+[1.4.4]: https://github.com/tokend/new-js-sdk/compare/1.4.3...1.4.4
 [1.4.3]: https://github.com/tokend/new-js-sdk/compare/1.4.2...1.4.3
 [1.4.2]: https://github.com/tokend/new-js-sdk/compare/1.4.1...1.4.2
 [1.4.1]: https://github.com/tokend/new-js-sdk/compare/1.4.0...1.4.1